backend/requirements.txt
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9
SQLAlchemy>=2.0.30
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
celery>=5.3.6
redis>=5.0.4
langgraph>=0.2.30
langchain-core>=0.2.38
python-dotenv>=1.0.1
structlog>=24.1.0
descope>=1.0.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
requests>=2.31.0
openai>=1.0.0
itsdangerous>=2.1.2